Revert "Reafactor/94 refactor"
