--- src/main/java/team/router/recycle/domain/station/StationRedisService.java
@@ -1,9 +1,0 @@
-package team.router.recycle.domain.station;
-
-
-public interface StationRedisService {
-
-    boolean isValid(String stationId);
-
-    boolean isInvalid(String stationId);
-}
